[Feat] 셀 분리 (#3)
